Split outreach into app/view structure
