+++ src/js/outreach/entities/index.js
@@ -1,0 +1,65 @@
+import Radio from 'backbone.radio';
+import fetcher, { handleJSON } from 'js/base/fetch';
+import { v4 as uuid } from 'uuid';
+
+function getRelationship(id, type) {
+  return { data: { type, id } };
+}
+
+let token;
+
+Radio.reply('auth', {
+  setToken(tokenString) {
+    token = tokenString;
+  },
+  getToken() {
+    return token;
+  },
+});
+
+function getToken({ dob, actionId }) {
+  const data = {
+    type: 'patient-tokens',
+    id: uuid(),
+    attributes: {
+      reason: 'outreach',
+      birth_date: dob,
+    },
+    relationships: {
+      action: getRelationship(actionId, 'patient-actions'),
+    },
+  };
+
+  return fetcher('/api/patient-tokens', {
+    method: 'POST',
+    data: JSON.stringify({ data }),
+  })
+    .then(handleJSON)
+    .then(({ data: { attributes } }) => {
+      Radio.request('auth', 'setToken', attributes.token);
+      return Promise.resolve(attributes.token);
+    });
+}
+
+function postResponse({ formId, actionId, response }) {
+  const data = {
+    type: 'form-responses',
+    id: uuid(),
+    attributes: { response },
+    relationships: {
+      action: getRelationship(actionId, 'patient-actions'),
+      form: getRelationship(formId, 'forms'),
+    },
+  };
+
+  return fetcher(`/api/actions/${ actionId }/relationships/form-responses`, {
+    method: 'POST',
+    data: JSON.stringify({ data }),
+  })
+    .then(handleJSON);
+}
+
+export {
+  getToken,
+  postResponse,
+};
