+++ src/js/outreach/views/form.scss
@@ -1,0 +1,27 @@
+
+.form__header {
+  align-items: center;
+  background: #F2F9FD;
+  display: flex;
+  font-size: 16px;
+  font-weight: 600;
+  line-height: 48px;
+  padding: 10px 24px;
+}
+
+.form__title {
+  flex-grow: 1;
+  line-height: 24px;
+  text-align: left;
+}
+
+.form__content {
+  background-color: $white;
+  height: calc(100vh - 68px);
+
+  iframe {
+    display: block;
+    height: 100%;
+    width: 100%;
+  }
+}
