+++ src/js/outreach/views/form_views.js
@@ -1,0 +1,46 @@
+import hbs from 'handlebars-inline-precompile';
+import { View } from 'marionette';
+
+import IframeFormBehavior from 'js/behaviors/iframe-form';
+
+import 'scss/modules/buttons.scss';
+import './form.scss';
+
+const iFrameFormView = View.extend({
+  behaviors: [IframeFormBehavior],
+  regions: {
+    formAction: '[data-action-region]',
+  },
+  template: hbs`
+  <div class="form__header">
+    <div class="form__title">{{ name }}</div>
+    <div data-action-region></div>
+  </div>
+  <div class="form__content">
+    <iframe src="/formapp/"></iframe>
+  </div>
+  `,
+});
+
+const SaveView = View.extend({
+  isDisabled: false,
+  tagName: 'button',
+  className: 'button--green',
+  attributes() {
+    return {
+      disabled: this.getOption('isDisabled'),
+    };
+  },
+  template: hbs`Save`,
+  triggers: {
+    'click': 'click',
+  },
+  onClick() {
+    this.$el.prop('disabled', true);
+  },
+});
+
+export {
+  iFrameFormView,
+  SaveView,
+};
