+++ src/js/outreach/views/login_views.js
@@ -1,65 +1,7 @@
 import hbs from 'handlebars-inline-precompile';
 import { View } from 'marionette';
 
-import IframeFormBehavior from 'js/behaviors/iframe-form';
-
-import { auth0Config } from 'js/config';
-
 import 'scss/modules/buttons.scss';
-import 'scss/outreach-core.scss';
-import './outreach.scss';
-
-const DialogView = View.extend({
-  template: hbs`
-    <div><h1 class="site-title">{{ name }}</h1></div>
-    <div class="dialog" data-content-region>
-      <div class="dialog__icon--success">{{fas "circle-check"}}</div>
-      <div>You’ve submitted the form. Nice job.</div>
-    </div>
-  `,
-  templateContext() {
-    return {
-      name: auth0Config.name,
-    };
-  },
-  regions: {
-    content: '[data-content-region]',
-  },
-});
-
-const iFrameFormView = View.extend({
-  behaviors: [IframeFormBehavior],
-  regions: {
-    formAction: '[data-action-region]',
-  },
-  template: hbs`
-  <div class="form__header">
-    <div class="form__title">{{ name }}</div>
-    <div data-action-region></div>
-  </div>
-  <div class="form__content">
-    <iframe src="/formapp/"></iframe>
-  </div>
-  `,
-});
-
-const SaveView = View.extend({
-  isDisabled: false,
-  tagName: 'button',
-  className: 'button--green',
-  attributes() {
-    return {
-      disabled: this.getOption('isDisabled'),
-    };
-  },
-  template: hbs`Save`,
-  triggers: {
-    'click': 'click',
-  },
-  onClick() {
-    this.$el.prop('disabled', true);
-  },
-});
 
 const LoginView = View.extend({
   ui: {
@@ -105,19 +47,8 @@ const NotAvailableView = View.extend({
   `,
 });
 
-const ErrorView = View.extend({
-  template: hbs`
-    <div class="dialog__icon--error">{{far "circle-exclamation"}}</div>
-    <div>Uh-oh, there was an error. Try reloading the page.</div>
-  `,
-});
-
 export {
-  DialogView,
-  iFrameFormView,
-  SaveView,
   LoginView,
   ResponseErrorView,
   NotAvailableView,
-  ErrorView,
 };
